# src/main/webapp/resources/rspace_api_specs_2_20_0.yaml
swagger: '2.0'
info:
  title: RSpace API
  description: >
    Welcome to the RSpace API. This API enables you to search for and retrieve
    your ELN data. It contains methods to create documents, upload files, and
    search, share and retrieve them.

    This documentation enables you to interactively explore the API using your
    own data, and to generate `curl` and client code snippets.


    API endpoints require client to be authenticated, so to use this documentation
    interactively you need to click on 'Authorize' button below and authenticate
    with your preferred method.

    #### Authentication with an API key

    To create an API key go to 'Manage API Key' section of your profile page
    (MyRSpace -> Profile). You should keep your API key private. You can
    regenerate API key at any time, which will invalidate the old key.

    #### Authentication with OAuth

    We have limited support for OAuth 'password' grant flow using `/oauth/token`
    endpoint. You can register OAuth app on your profile page (My RSpace -> Profile).
    For more information check our [Github documentation](https://github.com/rspace-os/api-tutorial/blob/master/oauth.md).


    Note: when using OAuth with this interactive documentation change 'type' option
    of 'Authorize' dialog from default 'Request body' to 'Query parameters'.

    #### Versioning notes

    The current _API_ version of the API is `version 1` and all URL paths begin
...
    unchanged.


...


...
    for details.

...

...

...

...
basePath: /api/v1
produces:
...
consumes:
...
security:
  - ApiKeySecurity: []
  - OAuth: []
securityDefinitions:
  ApiKeySecurity:
    description: Authentication system based on API Key.
    type: apiKey
    in: header
    name: apiKey
  OAuth:
    type: oauth2
    flow: password
    tokenUrl: /oauth/token
    scopes:
      all: Grants access to whole of API
paths:
...
  /stoichiometry/link:
    post:
      summary: Create a link between a stoichiometry molecule and an Inventory Container, Sample or Subsample
      tags:
        - Chemistry
      description: |
        Creates a link between a stoichiometry molecule and an Inventory item.
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/StoichiometryInventoryLinkRequest'
      responses:
        '201':
          description: Created link
          schema:
            $ref: '#/definitions/StoichiometryInventoryLinkDTO'
        '401':
          description: Authentication error or insufficient permissions
          schema:
            type: object
            properties:
              message:
                type: string
                example: "User does not have write permissions on document containing stoichiometry"
        '404':
          description: Stoichiometry molecule or inventory item not found
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Stoichiometry molecule or inventory item not found"
        '422':
          description: Unprocessable request. Examples include linking a molecule that already has a link, attempting to link to a Sample Template, or missing required quantity/unit values.
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Stoichiometry molecule already has an inventory link"
        '500':
          description: An unexpected error occurred while creating the link
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Error creating link"
    put:
      summary: Update quantity used for a stoichiometry-inventory link
      tags:
        - Chemistry
      description: |
        Updates the quantity used on a stoichiometry–inventory link.
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/StoichiometryLinkQuantityUpdateRequest'
      responses:
        '200':
          description: Updated link
          schema:
            $ref: '#/definitions/StoichiometryInventoryLinkDTO'
        '401':
          description: Authentication error or insufficient permissions
          schema:
            type: object
            properties:
              message:
                type: string
                example: "User does not have write permissions on document containing stoichiometry"
        '404':
          description: Link not found
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Link not found"
        '500':
          description: An unexpected error occurred while updating the link
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Error updating link"
  /stoichiometry/link/stock-deductions:
    post:
      summary: Batch deduct stock for a list of stoichiometry-inventory links
      tags:
        - Chemistry
      description: |
        Attempts to deduct stock for each stoichiometry-inventory link ID provided. This is a batch operation that can partially succeed.
        For each link, if it points to a SubSample and has not been fully deducted, the specified quantity is subtracted from the SubSample stock.
      parameters:
        - in: body
          name: linkIds
          description: List of stoichiometry-inventory link IDs to deduct stock for.
          required: true
          schema:
            type: array
            items:
              type: integer
              format: int64
      responses:
        '200':
          description: Batch operation result
          schema:
            $ref: '#/definitions/StoichiometryLinkStockReductionResult'
        '401':
          description: Authentication error
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Unauthorized"
        '500':
          description: Unexpected server error
          schema:
            type: object
            properties:
              message:
                type: string
                example: "An unexpected error occurred"
  /stoichiometry/link/{id}:
    get:
      summary: Get a stoichiometry-inventory link by id
      tags:
        - Chemistry
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: The link
          schema:
            $ref: '#/definitions/StoichiometryInventoryLinkDTO'
        '401':
          description: Authentication error or insufficient permissions
          schema:
            type: object
            properties:
              message:
                type: string
                example: "User does not have read permissions on document containing stoichiometry"
        '404':
          description: Link not found
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Link not found"
        '500':
          description: An unexpected error occurred while retrieving the link
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Error retrieving link"
    delete:
      summary: Delete a stoichiometry-inventory link
      tags:
        - Chemistry
      description: |
        Deletes the link. If stock was deducted earlier (when `stockDeducted=true`), deleting the link does not
        replenish stock.
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: Deletion succeeded
          schema:
            type: boolean
        '401':
          description: Authentication error or insufficient permissions
          schema:
            type: object
            properties:
              message:
                type: string
                example: "User does not have write permissions on document containing stoichiometry"
        '404':
          description: Link not found
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Link not found"
        '500':
          description: An unexpected error occurred while deleting the link
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Error deleting link"
...
definitions:
...
  QuantityInfo:
    description: Quantity information with numeric value and unit identifier.
    properties:
      numericValue:
        type: number
        format: double
        description: Numeric value of the quantity
        example: 10.5
      unitId:
        type: integer
        format: int32
        description: Unit identifier from RSUnitDef. Options can be retrieved from /units endpoint.
        example: 6
    example:
      numericValue: 10.5
      unitId: 6
  StoichiometryInventoryLinkDTO:
    description: Link between a stoichiometry molecule and an Inventory item
    properties:
      id:
        type: integer
        format: int64
        example: 1234
      inventoryItemGlobalId:
        type: string
        description: Global identifier of the inventory item (e.g. SA123, SS456, IC789)
        example: 'SS456'
      stoichiometryMoleculeId:
        type: integer
        format: int64
        example: 567
      quantity:
        $ref: '#/definitions/QuantityInfo'
      stockDeducted:
        type: boolean
        description: |-
          If true, stock has been previously deducted for this link.
        default: false
        example: false
    example:
      id: 1234
      inventoryItemGlobalId: 'SS456'
      stoichiometryMoleculeId: 567
      quantity:
        numericValue: 10.5
        unitId: 6
      stockDeducted: false
  StoichiometryInventoryLinkRequest:
    description: >-
      Request to create a link between a stoichiometry molecule and an Inventory item.
      Allowed inventory targets are Containers, Samples and Subsamples. Linking to Sample Templates is not permitted.
    required:
      - inventoryItemGlobalId
      - stoichiometryMoleculeId
      - quantity
      - unitId
    properties:
      inventoryItemGlobalId:
        type: string
        description: >-
          Global identifier of the target inventory item (allowed prefixes: IC for Containers, SA for Samples, SS for Subsamples; e.g. IC789, SA123, SS456). Sample Templates are not allowed.
        example: 'SS456'
      stoichiometryMoleculeId:
        type: integer
        format: int64
        example: 567
      quantity:
        type: number
        format: double
        description: Numeric value of the quantity used
        example: 10.5
      unitId:
        type: integer
        format: int32
        description: Unit identifier. Options can be retrieved from /units endpoint.
        example: 6
    example:
      inventoryItemGlobalId: 'SS456'
      stoichiometryMoleculeId: 567
      quantity: 10.5
      unitId: 6
  StoichiometryLinkQuantityUpdateRequest:
    description: Request to update quantity used for a stoichiometry-inventory link
    required:
      - stoichiometryLinkId
      - newQuantity
    properties:
      stoichiometryLinkId:
        type: integer
        format: int64
        example: 1234
      newQuantity:
        $ref: '#/definitions/QuantityInfo'
    example:
      stoichiometryLinkId: 1234
      newQuantity:
        numericValue: 15.0
        unitId: 6
  StoichiometryLinkStockReductionResult:
    description: Result of a batch stock reduction operation
    properties:
      results:
        type: array
        items:
          $ref: '#/definitions/StoichiometryLinkStockReductionIndividualResult'
    example:
      results:
        - linkId: 1234
          success: true
        - linkId: 1235
          success: false
          error:
            httpCode: 422
            internalCode: 42202
            message: "Insufficient stock for SubSample SS457"
  StoichiometryLinkStockReductionIndividualResult:
    description: Individual result for a single link ID in a batch reduction
    properties:
      linkId:
        type: integer
        format: int64
        example: 1234
      success:
        type: boolean
        example: true
      error:
        $ref: '#/definitions/ApiError'
    example:
      linkId: 1234
      success: true
  ApiError:
    description: Error details returned by the API
    properties:
      status:
        type: string
        description: HTTP status name
        example: UNPROCESSABLE_ENTITY
      httpCode:
        type: integer
        description: HTTP status code
        example: 422
      internalCode:
        type: integer
        description: RSpace internal error code
        example: 42201
      message:
        type: string
        description: Main error message
        example: "Insufficient stock for SubSample"
      messageCode:
        type: string
        description: Error message code for translation
        example: "error.inventory.insufficient.stock"
      errors:
        type: array
        items:
          type: string
    example:
      status: UNPROCESSABLE_ENTITY
      httpCode: 422
      internalCode: 42201
      message: "Insufficient stock for SubSample"
      messageCode: "error.inventory.insufficient.stock"
...